+++ course_settings.php
@@ -27,7 +27,7 @@
  **/
 
 require_once(__DIR__.'/../../../config.php');
-require_once('payment_settings_form.php');
+require_once('form/payment_settings_form.php');
 
 $courseid = optional_param('id', 0, PARAM_INT);
 if (empty($courseid)) {
